Ultimo Proyecto

+++ Assets/Scripts/Saves/JsonSaves.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using System.IO;
+using UnityEngine.SceneManagement;
 
 public class JsonSaves : MonoBehaviour
 {
@@ -21,7 +22,7 @@ public class JsonSaves : MonoBehaviour
 
         _path = customDir + "/Saves.Json";
 
-        Debug.Log(_path);
+        Debug.LogWarning(_path);
     }
 
     private void OnApplicationQuit()
@@ -90,11 +91,18 @@ public class JsonSaves : MonoBehaviour
 
     public void DeleteJSON()
     {
-        Debug.Log("Se borro el save data");
-        File.Delete(_path);
+        Debug.LogWarningFormat("Se borro el save data");
 
-        CallJson.instance.save.GetSaveData.tutorialCompletedJSON = false;
-        CallJson.instance.save.GetSaveData.moneyJSON = 0;
+        if(File.Exists(_path))
+        {
+            File.Delete(_path);
+
+            CallJson.instance.save.GetSaveData.tutorialCompletedJSON = false;
+            CallJson.instance.save.GetSaveData.moneyJSON = 0;
+
+            SceneManager.LoadScene(1);
+        }
+
 
         //saveData = new SaveData();//Si quiero resetear los datos
     }
